backend/requirements.txt
flask
flask-cors
llama-cpp-python==0.3.12
openai
python-dotenv
tqdm
numpy
